move code only used by cloud to cloud project
